feat(cmake): Add option to install DLL
This commit introduces an option to install the DLL into the game directory.

The `SHOULD_INSTALL_DLL` option is added to `CMakeLists.txt`.  If `SHOULD_INSTALL_DLL` is on, the DLL is copied to the install directory; otherwise, it's skipped.  The build process now conditionally copies the DLL.  The install target now depends on the DLL copy command only if the option is set. This improves build flexibility.